requirements.txt
web3==7.6.0
python-dotenv==1.0.0
hexbytes~=0.3.0
aiohttp~=3.8.4
requests==2.31.0